NodeGenerator script now instantiates a node prefab
The NodeGenerator script was modified to allow the NodeGenerator instantiate a prefab that represents the position of the nodes, so in this way it's easier to click and watch descriptions of an specific node. Also it is useful for making a more accurate track of the red box movement with our eyes.