Path finding and FSM merged
Tha path finding system and the FSM were merged to work with the miner simulation. Also a new event was added to the FSM that makes the miner dont go again to the mine if there is no gold in it, so in this way the miner jumps to the "stay still" state when the proper checking gives true as result when the FSM is in the "leave gold" state.